类型宽度收口：别名调色板补齐 + 宽度显式化 + StepUtility 补重载与 ParseInteger
本提交含两批工作，二者改动落在同一批文件里，按文件切会切出不能编译的中间态。

别名调色板（配合 Templates 仓库）：
- Types.h 由模板重生成：新增 UInt8/Int8/Int16/UInt32/UInt64 别名，宽度断言保留
- StepUtility 补 5 个 WriteString 重载。不补时新族会落到末尾那个 %s 模板重载上，
  sprintf("%d=%s", key, value) 会把整型当 char* 解引用，是必崩路径
- 新增 ParseInteger（std::from_chars）：格式非法或越界都返回 false，
  供生成物里的窄整型反序列化做范围校验，取代静默截断的 atoi
- 生成物与手写源码里的 UShortType/IntType 全部改名（Types.xml 已无这两个别名）

类型宽度显式化：
- Types.h 的别名宽度从"编译器把 int 当几位"改为定宽（unsigned short -> uint16_t、
  int -> int32_t）；64 位保留 long long，因为 Linux 上 int64_t 是 long，
  改它会让全仓 %lld 变成格式不匹配
- ProtocolVersion.h 补 HeadField/TailField 的字段偏移断言：只钉总长 16/4 会漏掉
  "字段重排但总长不变"
- PackageReader.cpp 三处 %u 配 UShortType 改 static_cast<unsigned int>：
  窄整型进可变参数会先升格成 int，%u 在标准上不成立（GCC -Wformat 会报）

实证：cmake --build out/build/x64-Debug 零错误；bin/Debug/UnitTests.exe
366/366 通过（25 套件）。

未覆盖：新族当前没有任何模型在用，故新的反序列化分支与 5 个重载在现有用例里
不被执行——366 与加调色板之前是同一个数字，不能当作新路径已生效的证据。

+++ src/Network/Protocol/PackageReader.cpp
@@ -90,7 +90,7 @@ bool PackageReader::IsBodyLenWithinFrameLimit() const
 		return true;
 	}
 	WriteLog(LogLevel::Warning, "Body Length Exceeds Frame Limit. BodyLen:%u, MaxFrameBodyLen:%u, SessionID:%lld, IP:%s",
-		m_Head.BodyLen, MaxFrameBodyLen, m_SessionID, m_IPAddress);
+		static_cast<unsigned int>(m_Head.BodyLen), MaxFrameBodyLen, m_SessionID, m_IPAddress);
 	return false;
 }
 PackageReader::AlignResult PackageReader::AlignToAnchor(const char* anchor, unsigned int anchorLength)
@@ -156,15 +156,16 @@ bool PackageReader::ParseXtpPackage(Package*& package)
 		if (m_Head.Version != ProtocolVersionValue)
 		{
 			WriteLog(LogLevel::Error, "Protocol Version Not Match. RemoteVersion:%u, LocalVersion:%u, SessionID:%lld, IP:%s",
-				m_Head.Version, ProtocolVersionValue, m_SessionID, m_IPAddress);
+				static_cast<unsigned int>(m_Head.Version), static_cast<unsigned int>(ProtocolVersionValue),
+				m_SessionID, m_IPAddress);
 			return false;
 		}
 		if (!IsBodyLenWithinFrameLimit())
-	{
-		DiscardFront(1);
-		continue;
-	}
-	if (m_Length < (sizeof(HeadField) + m_Head.BodyLen + sizeof(TailField)))
+		{
+			DiscardFront(1);
+			continue;
+		}
+		if (m_Length < (sizeof(HeadField) + m_Head.BodyLen + sizeof(TailField)))
 		{
 			return true;
 		}
@@ -236,15 +237,16 @@ bool PackageReader::ParseStepPackage(Package*& package)
 		if (m_Head.Version != ProtocolVersionValue)
 		{
 			WriteLog(LogLevel::Error, "Protocol Version Not Match. RemoteVersion:%u, LocalVersion:%u, SessionID:%lld, IP:%s",
-				m_Head.Version, ProtocolVersionValue, m_SessionID, m_IPAddress);
+				static_cast<unsigned int>(m_Head.Version), static_cast<unsigned int>(ProtocolVersionValue),
+				m_SessionID, m_IPAddress);
 			return false;
 		}
 		if (!IsBodyLenWithinFrameLimit())
-	{
-		DiscardFront(1);
-		continue;
-	}
-	int tailIndex = headEndIndex + m_Head.BodyLen;
+		{
+			DiscardFront(1);
+			continue;
+		}
+		int tailIndex = headEndIndex + m_Head.BodyLen;
 		if (m_Length < unsigned(tailIndex + StepTailLen))
 		{
 			return true;
